Extract ChangelogView from main application controller (#477)
- add `changelog-view.fxml`
- add `ChangelogViewController` and move (old) code related to compute
and display the changelog
- introduce a property with binding to hook up the changelog update

Contributes to #464.
Part of #462
