+++ src/main/java/org/terasology/launcher/gui/javafx/ApplicationController.java
@@ -35,10 +35,8 @@ import javafx.scene.control.ButtonType;
 import javafx.scene.control.ComboBox;
 import javafx.scene.control.ProgressBar;
 import javafx.scene.control.Tooltip;
-import javafx.scene.effect.BlendMode;
 import javafx.scene.image.ImageView;
 import javafx.scene.input.MouseEvent;
-import javafx.scene.web.WebView;
 import javafx.stage.Modality;
 import javafx.stage.Stage;
 import javafx.stage.StageStyle;
@@ -170,8 +168,6 @@ public class ApplicationController {
     @FXML
     private Button deleteButton;
     @FXML
-    private WebView changelogView;
-    @FXML
     private Button settingsButton;
     @FXML
     private Button exitButton;
@@ -187,9 +183,13 @@ public class ApplicationController {
     @FXML
     private LogViewController logViewController;
     @FXML
+    private ChangelogViewController changelogViewController;
+    @FXML
     private FooterController footerController;
 
-    /** Indicate whether the user's hard drive is running out of space for game downloads. */
+    /**
+     * Indicate whether the user's hard drive is running out of space for game downloads.
+     */
     final private Property<Optional<Warning>> warning;
 
     public ApplicationController() {
@@ -565,7 +565,6 @@ public class ApplicationController {
     private void updateGui() {
         updateButtons();
         updateTooltipTexts();
-        updateChangeLog();
     }
 
     private void updateButtons() {
@@ -596,7 +595,6 @@ public class ApplicationController {
             cancelDownloadButton.setVisible(false);
             startAndDownloadButton.setVisible(true);
         }
-
     }
 
     private void updateJobBox() {
@@ -639,114 +637,6 @@ public class ApplicationController {
 //                buildVersionBox.getSelectionModel().select(versionItem);
 //            }
 //        }
-    }
-
-    private void updateChangeLog() {
-        final TerasologyGameVersion gameVersion = getSelectedGameVersion();
-        final String gameInfoTextHTML;
-        if ((gameVersion == null) || (gameVersion.getJob() == null) || (gameVersion.getBuildNumber() == null)) {
-            gameInfoTextHTML = "";
-        } else {
-            gameInfoTextHTML = getGameInfoText(gameVersion);
-        }
-
-        changelogView.getEngine().loadContent(gameInfoTextHTML);
-        changelogView.setBlendMode(BlendMode.LIGHTEN);
-        changelogView.getEngine().setUserStyleSheetLocation(BundleUtils.getFXMLUrl("css_webview").toExternalForm());
-    }
-
-    private String getGameInfoText(TerasologyGameVersion gameVersion) {
-//        logger.debug("Display game version: {} {}", gameVersion, gameVersion.getGameVersionInfo());
-//
-//        final Object[] arguments = new Object[9];
-//        arguments[0] = gameVersion.getJob().name();
-//        if (gameVersion.getJob().isStable()) {
-//            arguments[1] = 1;
-//        } else {
-//            arguments[1] = 0;
-//        }
-//        arguments[2] = gameVersion.getJob().getGitBranch();
-//        arguments[3] = gameVersion.getBuildNumber();
-//        if (gameVersion.isLatest()) {
-//            arguments[4] = 1;
-//        } else {
-//            arguments[4] = 0;
-//        }
-//        if (gameVersion.isInstalled()) {
-//            arguments[5] = 1;
-//        } else {
-//            arguments[5] = 0;
-//        }
-//        if (gameVersion.getSuccessful() != null) {
-//            if (!gameVersion.getSuccessful()) {
-//                // faulty
-//                arguments[6] = 0;
-//            } else {
-//                arguments[6] = 1;
-//            }
-//        } else {
-//            // unknown
-//            arguments[6] = 2;
-//        }
-//        if ((gameVersion.getGameVersionInfo() != null)
-//                && (gameVersion.getGameVersionInfo().getDisplayVersion() != null)) {
-//            arguments[7] = gameVersion.getGameVersionInfo().getDisplayVersion();
-//        } else {
-//            arguments[7] = "";
-//        }
-//        if ((gameVersion.getGameVersionInfo() != null)
-//                && (gameVersion.getGameVersionInfo().getDateTime() != null)) {
-//            arguments[8] = gameVersion.getGameVersionInfo().getDateTime();
-//        } else {
-//            arguments[8] = "";
-//        }
-//
-//        final String infoHeader1 = BundleUtils.getMessage(gameVersion.getJob().getInfoMessageKey(), arguments);
-//        final String infoHeader2 = BundleUtils.getMessage("infoHeader2", arguments);
-//
-//        final StringBuilder b = new StringBuilder();
-//        if ((infoHeader1 != null) && (infoHeader1.trim().length() > 0)) {
-//            b.append("<h1>")
-//                    .append(escapeHtml(infoHeader1))
-//                    .append("</h1>\n");
-//        }
-//        if ((infoHeader2 != null) && (infoHeader2.trim().length() > 0)) {
-//            b.append("<h2>")
-//                    .append(escapeHtml(infoHeader2))
-//                    .append("</h2>\n");
-//        }
-//        b.append("<strong>\n")
-//                .append(BundleUtils.getLabel("infoHeader3"))
-//                .append("</strong>\n");
-//
-//        if ((gameVersion.getChangeLog() != null) && !gameVersion.getChangeLog().isEmpty()) {
-//            b.append("<p>\n")
-//                    .append(BundleUtils.getLabel("infoHeader4"))
-//                    .append("<ul>\n");
-//            for (String msg : gameVersion.getChangeLog()) {
-//                b.append("<li>")
-//                        .append(escapeHtml(msg))
-//                        .append("</li>\n");
-//            }
-//            b.append("</ul>\n")
-//                    .append("</p>\n");
-//        }
-//
-//        /* Append changelogs of previous builds. */
-//        int previousLogs = gameVersion.getJob().isStable() ? 1 : 10;
-//        b.append("<hr/>");
-//        for (String msg : packageManager.getAggregatedChangeLog(gameVersion, previousLogs)) {
-//            b.append("<li>")
-//                    .append(escapeHtml(msg))
-//                    .append("</li>\n");
-//        }
-//
-//        return b.toString();
-        return "WIP";
-    }
-
-    private static String escapeHtml(String text) {
-        return text.replace("&", "&amp;").replace("<", "&lt;").replace(">", "&gt;").replace("\"", "&quot;").replace("'", "&#x27;").replace("/", "&#x2F;");
     }
 
     void finishedGameDownload(boolean cancelled, boolean successfulDownloadAndExtract, boolean successfulLoadVersion, Path gameDirectory) {
